api/requirements.txt
cornice
cornice_swagger
pyramid_chameleon
passlib
waitress
sqlalchemy
sqlalchemy-utils
pyramid_tm
pyramid_retry
pyramid_jwt
zope.sqlalchemy
psycopg2-binary
psycopg2
gunicorn
intervals
google-cloud-storage
google-cloud-bigquery
google-cloud-logging
grpcio
requests
